+++ packages/pn-commons/src/components/NotificationDetail/__test__/NotificationDetailTimeline.test.tsx
@@ -1,15 +1,15 @@
-import { fireEvent, waitFor, within, screen } from '@testing-library/react';
+import { fireEvent, waitFor, screen, prettyDOM } from '@testing-library/react';
 
 import { render } from '../../../test-utils';
-import { getNotificationStatusLabelAndColorFromTimelineCategory } from '../../../utils/status.utility';
 import { getDay, getMonthString, getTime } from '../../../utils/date.utility';
 import {
   INotificationDetailTimeline,
   NotificationStatusHistory,
   TimelineCategory,
-} from '../../../types/Notifications';
+} from '../../../types/NotificationDetail';
 import { NotificationStatus } from '../../../types/NotificationStatus';
-import * as hooks from "../../../hooks/IsMobile.hook";
+import * as hooks from '../../../hooks/IsMobile.hook';
+import { getNotificationStatusLabelAndColor } from '../../../utils/status.utility';
 import NotificationDetailTimeline from '../NotificationDetailTimeline';
 
 const timeline: Array<INotificationDetailTimeline> = [
@@ -30,48 +30,49 @@ const timeline: Array<INotificationDetailTimeline> = [
       category: TimelineCategory.END_OF_ANALOG_DELIVERY_WORKFLOW,
       taxdId: '',
     },
-  }
+  },
 ];
 
 const statusHistory: Array<NotificationStatusHistory> = [
   {
     status: NotificationStatus.ACCEPTED,
-    activeFrom: '',
-    relatedTimelineElements: ['mocked-timeline-id-1'],
+    activeFrom: '2022-03-21T08:56:50.177Z',
+    relatedTimelineElements: ['mocked-id-1'],
   },
   {
     status: NotificationStatus.DELIVERED,
-    activeFrom: '',
-    relatedTimelineElements: ['mocked-timeline-id-2'],
-  }
+    activeFrom: '2022-01-15T08:56:50.177Z',
+    relatedTimelineElements: ['mocked-id-2'],
+  },
 ];
 
 const useIsMobileSpy = jest.spyOn(hooks, 'useIsMobile');
 
 const testTimelineRendering = async (container: HTMLElement) => {
-  const timelineItems = container.querySelectorAll('[data-testid="timelineItem"]');
-    expect(timelineItems).toHaveLength(timeline.length);
-    // beacuse of await into loop, we need to use for loop and not forEach
-    let timelineIndex = 0;
-    for (const item of timelineItems) {
-      const dateItems = item.querySelectorAll('p');
+  const timelineItems = container.querySelectorAll('.MuiTimelineItem-root');
+  // we multiply for 2 because, for each status history element, there is two timeline elements (status + moreLessButton)
+  expect(timelineItems).toHaveLength(statusHistory.length * 2);
+  timelineItems.forEach((item, timelineIndex) => {
+    if (timelineIndex % 2 === 0) {
+      const dateItems = item.querySelectorAll('[data-testid="dateItem"]');
       expect(dateItems).toHaveLength(3);
-      expect(dateItems[0]).toHaveTextContent(getMonthString(timeline[timelineIndex].timestamp));
-      expect(dateItems[1]).toHaveTextContent(getDay(timeline[timelineIndex].timestamp));
-      expect(dateItems[2]).toHaveTextContent(getTime(timeline[timelineIndex].timestamp));
-      const itemStatus = await within(item).findByTestId('itemStatus');
+      expect(dateItems[0]).toHaveTextContent(getMonthString(statusHistory[timelineIndex / 2].activeFrom));
+      expect(dateItems[1]).toHaveTextContent(getDay(statusHistory[timelineIndex / 2].activeFrom));
+      expect(dateItems[2]).toHaveTextContent(getTime(statusHistory[timelineIndex / 2].activeFrom));
+      const itemStatus = item.querySelector('[data-testid="itemStatus"]');
       expect(itemStatus).toBeInTheDocument();
       const classRoot = 'MuiChip-color';
-      const { color, label } = getNotificationStatusLabelAndColorFromTimelineCategory(
-        timeline[timelineIndex],
-        statusHistory
-      );
+      const { color, label } = getNotificationStatusLabelAndColor(statusHistory[timelineIndex / 2].status);
       const buttonClass = `${classRoot}${color!.charAt(0).toUpperCase() + color!.slice(1)}`;
       expect(itemStatus).toHaveTextContent(label);
-      expect(itemStatus.classList.contains(buttonClass)).toBe(true);
-      timelineIndex++;
+      expect(itemStatus!.classList.contains(buttonClass)).toBe(true);
+    } else {
+      const moreLessButton = item.querySelector('[data-testid="moreLessButton"]');
+      expect(moreLessButton).toBeInTheDocument();
+      expect(moreLessButton).toHaveTextContent(/mocked-more-label/i);
     }
-}
+  });
+};
 
 describe('NotificationDetailTimeline Component', () => {
   it('renders NotificationDetailTimeline (desktop)', async () => {
@@ -84,7 +85,9 @@ describe('NotificationDetailTimeline Component', () => {
         statusHistory={statusHistory}
         clickHandler={jest.fn()}
         legalFactLabel="mocked-legalFact-label"
-        historyButtonLabel='mocked-history-label'
+        historyButtonLabel="mocked-history-label"
+        showLessButtonLabel="mocked-less-label"
+        showMoreButtonLabel="mocked-more-label"
       />
     );
     expect(result?.container).toHaveTextContent(/mocked-title/i);
@@ -102,11 +105,13 @@ describe('NotificationDetailTimeline Component', () => {
         statusHistory={statusHistory}
         clickHandler={jest.fn()}
         legalFactLabel="mocked-legalFact-label"
-        historyButtonLabel='mocked-history-label'
+        historyButtonLabel="mocked-history-label"
+        showLessButtonLabel="mocked-less-label"
+        showMoreButtonLabel="mocked-more-label"
       />
     );
     expect(result?.container).toHaveTextContent(/mocked-title/i);
-    const timelineItems = await result?.findAllByTestId('timelineItem');
+    const timelineItems = result?.container.querySelectorAll('.MuiTimelineItem-root');
     expect(timelineItems).toHaveLength(1);
     const historyButton = await result?.findByTestId('historyButton');
     expect(historyButton!).toBeInTheDocument();
@@ -116,5 +121,35 @@ describe('NotificationDetailTimeline Component', () => {
     });
     expect(drawer!).toBeInTheDocument();
     await testTimelineRendering(drawer!);
+  });
+
+  it('expand timeline item (desktop)', async () => {
+    useIsMobileSpy.mockReturnValue(false);
+    // render component
+    const result = render(
+      <NotificationDetailTimeline
+        title="mocked-title"
+        timeline={timeline}
+        statusHistory={statusHistory}
+        clickHandler={jest.fn()}
+        legalFactLabel="mocked-legalFact-label"
+        historyButtonLabel="mocked-history-label"
+        showLessButtonLabel="mocked-less-label"
+        showMoreButtonLabel="mocked-more-label"
+      />
+    );
+    // get first moreLessButton
+    const moreLessButton = result?.container.querySelector('[data-testid="moreLessButton"] button');
+    fireEvent.click(moreLessButton!);
+    await waitFor(() => {
+      expect(moreLessButton).toHaveTextContent(/mocked-less-label/i);
+      const timelineExpandedItem = result?.container.querySelector('.MuiTimelineItem-root:nth-child(3)');
+      const dateItems = timelineExpandedItem!.querySelectorAll('[data-testid="dateItem"]');
+      expect(dateItems).toHaveLength(3);
+      expect(dateItems[0]).toHaveTextContent(getMonthString(timeline[0].timestamp));
+      expect(dateItems[1]).toHaveTextContent(getDay(timeline[0].timestamp));
+      expect(dateItems[2]).toHaveTextContent(getTime(timeline[0].timestamp));
+      expect(timelineExpandedItem).toHaveTextContent(timeline[0].category);
+    })
   });
 });
