--- packages/pn-commons/src/types/NotificationDetailTable.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from "react";
-
-export interface DetailTableRow {
-  id: number;
-  label: string;
-  value: ReactNode;
-}
